EMP-553: refactored crm4 models to be in their own package
