+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/mapper/Crm4Mapper.java
@@ -7,10 +7,10 @@ import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.Crm4Detail
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.Crm4FormDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.Crm4RelatedSubmissionDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.CrmStandardPropertiesDTO;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.Crm4AuthorisedExpenditureModel;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.Crm4DetailsModel;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.Crm4Model;
-import uk.gov.justice.laa.crime.equinity.historicaldata.model.Crm4SubmissionModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm4.Crm4AuthorisedExpenditureModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm4.Crm4DetailsModel;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm4.Crm4Model;
+import uk.gov.justice.laa.crime.equinity.historicaldata.model.crm4.Crm4SubmissionModel;
 
 @Mapper(componentModel = "spring")
 public interface Crm4Mapper extends CrmEvidenceFilesMapper {
