+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/crm4/Crm4AuthorisedExpenditureModel.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.equinity.historicaldata.model;
+package uk.gov.justice.laa.crime.equinity.historicaldata.model.crm4;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
